requirements.txt
bootstrap-admin==0.4.0
Django==2.1.5
djangorestframework==3.9.0
Pygments==2.3.1
pytz==2018.9
